#3 Weighted pattern implemented
Implemented the weighted pattern class, able to add rows to itself, retrieve weights, and calculate a score if given a certain pattern.

Also added a Data Entry class, a temporary class for representing data. Will be expanded upon in #5.

The function for generating rows is also incomplete, as that will be continued during Sprint 2