#2 Table can be filled and has a scroll bar
The table can now be filled with data entries, on which it calls the basic_info() function to get display info. The number of entries can exceed the space given on the table, and to accommodate this, a scroll bar has been implemented, along with click and mouse wheel interaction.

UI elements no longer stored in a list, but in a dict with a unique name, so that it can be accessed more easily from the outside.