#6 #9 SearchBox implemented
Implemented the SearchBox UI element, which combines a TextBox, Button, and Table into one single element. This element can take a user query and, on the press of the button, display the results of the search in the table.

Fixed some divide by zero errors in the scroll bar code.

Added some commentary to some functions.