#7 InfoScene basic structure implemented
Implemented the basic structure of the InfoScene. Entries inside a table now have another button. When pressed, a new overlay window is presented on which detailed information will be displayed. The InfoScene also contains a return button to remove the overlay and to return to the previous scene without interruption.