#7 Display title, cast and directors of a movie in the InfoScene
Cleaned the entry classes a bit, adding some more info and adding the ability to update an entry to collect more information from IMDbPy. Most of this information (for movies) is now presented on the InfoScene.

Also simplified how text is printed, generalizing it to a single function text(), instead of calling the renderer each time and blitting to the surface manually.